update parser and search services for handling the new format served from api
